Run game from GameRunner
